CI: Verify semver compatibility when tags is pushed
Add a job that checks compatibility when a tag is pushed.
The command executed is just
`cargo semver-checks -p scylla -p scylla-cql` because it's behavior is
exactly what we want. It gets current version from Cargo.toml,
calculates previous version (unless current version is on crates.io -
which should not be the case) and checks that the release type
(major / minor / patch) is correct.